# sdk/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "QuickLink API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/links/{shortCode}": {
      "get": {
        "operationId": "getLink",
        "parameters": [{ "name": "shortCode", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Link detail", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ShortLinkDetail" } } } } }
      },
      "patch": {
        "operationId": "updateLink",
        "parameters": [{ "name": "shortCode", "in": "path", "required": true, "schema": { "type": "string" } }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UpdateLinkInput" } } } },
        "responses": { "200": { "description": "Updated link", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ShortLinkDetail" } } } } }
      },
      "delete": {
        "operationId": "deleteLink",
        "parameters": [{ "name": "shortCode", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Deleted link", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/DeleteLinkResult" } } } } }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ShortLinkDetail": {
        "type": "object",
        "properties": {
          "id": { "type": "string" }, "originalUrl": { "type": "string" }, "shortCode": { "type": "string" },
          "title": { "type": ["string", "null"] }, "description": { "type": ["string", "null"] },
          "ogImage": { "type": ["string", "null"] }, "tags": { "type": "array", "items": { "type": "string" } },
          "clicks": { "type": "integer" }, "createdAt": { "type": "string" }, "updatedAt": { "type": "string" },
          "expiresAt": { "type": ["string", "null"] }, "expiredUrl": { "type": ["string", "null"] },
          "maxClicks": { "type": ["integer", "null"] }, "isActive": { "type": "boolean" },
          "cloaked": { "type": "boolean" }, "webhookUrl": { "type": ["string", "null"] },
          "metaPixelId": { "type": ["string", "null"] }, "googleTagId": { "type": ["string", "null"] },
          "xPixelId": { "type": ["string", "null"] }, "riskStatus": { "type": "string" },
          "healthStatus": { "type": "string" }
        },
        "required": ["id","originalUrl","shortCode"]
      },
...
      "UpdateLinkInput": {
        "type": "object",
        "properties": {
          "title": { "type": ["string", "null"] }, "description": { "type": ["string", "null"] },
          "ogImage": { "type": ["string", "null"] }, "isActive": { "type": "boolean" },
          "tags": { "type": "array", "items": { "type": "string" } },
          "expiresAt": { "type": ["string", "null"] }, "expiredUrl": { "type": ["string", "null"] },
          "password": { "type": ["string", "null"] }, "destinationUrl": { "type": "string" },
          "maxClicks": { "type": ["integer", "null"] }, "cloaked": { "type": "boolean" },
          "webhookUrl": { "type": ["string", "null"] }
        }
      },
...
      "DeleteLinkResult": {
        "type": "object",
        "properties": { "deleted": { "type": "boolean" }, "shortCode": { "type": "string" } },
        "required": ["deleted","shortCode"]
      },
...
    }
  }
}